Put generated code in memory:
